requirements.txt
flask==3.1.0
openai==1.70.0
python-dotenv==1.1.0
stripe==11.6.0
pdfkit==1.0.0
gunicorn==21.2.0
Jinja2==3.1.6
Werkzeug==3.1.3
